Api/requirements.txt
asgiref==3.9.2
Django==4.2.25
django-rest-framework==0.1.0
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
requests==2.31.0
django-cors-headers==4.3.1
